+++ tests/testthat/test-misc.R
@@ -1,12 +1,5 @@
 context("misc.R")
 
-test_that("`like` works", {
-  expect_true(suppressWarnings("test" %like% c("^t", "^s")))
-  expect_true("test" %like% "test")
-  expect_true("test" %like% "TEST")
-  expect_true(as.factor("test") %like% "TEST")
-})
-
 test_that("percentages works", {
   expect_equal(percent(0.25), "25%")
   expect_equal(percent(0.5), "50%")